pyproject.toml
[tool.poetry]
name = "web-app"
version = "0.1.0"
description = ""
authors = ["yosep34 <yosepalexsander9@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.1.2"
pandas = "^1.1.3"
numpy = "1.18.5"
djangorestframework = "^3.12.1"
django-cors-headers = "^3.5.0"
gunicorn = "^20.0.4"
django-heroku = "^0.3.1"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.6"
django-environ = "^0.4.5"
tensorflow = "2.3.0"
h5py = "^2.10.0"
django-bootstrap4 = "^2.3.1"
whitenoise = "5.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.4"
black = "^20.8b1"
pip = "^20.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
